0.4.1

+++ src/lib/common/update-file.ts
@@ -1,7 +1,7 @@
-import {MP3Reader, MP3ReaderOptions} from '../mp3/mp3_reader';
+import {MP3Reader, MP3ReaderOptions} from '../mp3/mp3.reader';
 import fse from 'fs-extra';
-import {FileWriterStream} from './streams';
 import {IMP3} from '../..';
+import {FileWriterStream} from './stream-writer-file';
 
 export async function updateFile(
 	filename: string, options: MP3ReaderOptions, keepBackup: boolean,
@@ -15,7 +15,7 @@ export async function updateFile(
 	}
 	const tmpFile = filename + '.tempmp3';
 	const bakFile = filename + '.bak';
-	let exists = await fse.pathExists(tmpFile);
+	const exists = await fse.pathExists(tmpFile);
 	if (exists) {
 		await fse.remove(tmpFile);
 	}
@@ -28,22 +28,19 @@ export async function updateFile(
 		return Promise.reject(e);
 	}
 	await fileWriterStream.close();
-	exists = await fse.pathExists(bakFile);
+	const bakExists = await fse.pathExists(bakFile);
 	if (keepBackup) {
-		if (!exists) {
+		if (!bakExists) {
 			await fse.rename(filename, bakFile);
 		} else {
 			// we have already a .bak which will be not touched
 			await fse.remove(filename);
 		}
-	} else {
-		if (exists) {
-			await fse.remove(bakFile);
-		}
+	} else if (!bakExists) {
 		await fse.rename(filename, bakFile);
 	}
 	await fse.rename(tmpFile, filename);
-	if (!keepBackup) {
+	if (!keepBackup && !bakExists) {
 		await fse.remove(bakFile);
 	}
 }
